ScoreMeter: Add "potential" declarer points
Instead of just displaying the lower limit of cards we know for sure the
declarer has, draw-in the potential points, that is what is possibly in
the dropped cards (at most 22 points).

We also account for the value of cards each player has on hand and what
is currently in the trick.

We shade the potential area a little less opaque.