DeclaringWidget: Recalculate on skat pick up
We need to make sure to recalculate the calculator data and reset the
form when the user picks up the skat.

This gives the <Calculator/> component a key dependend on
bidding.skatPickedUp, so the key changes when and recreates the
Form, repopulating it with fresh data.

Since that data comes from the redux state, we need to consider
the condition when updating a channel, too.